+++ README.md
@@ -49,5 +49,5 @@
 # Step4
 
 ## 요구사항
-- [ ] In -> Out 방식으로 TDD를 도전해 본다.
-- [ ] 인터페이스 기반으로 구현해 본다.
+- [x] In -> Out 방식으로 TDD를 도전해 본다.
+- [x] 인터페이스 기반으로 구현해 본다.
