+++ src/main/java/nextstep/ladder/domain/dimension/OneDimension.java
@@ -1,0 +1,5 @@
+package nextstep.ladder.domain.dimension;
+
+public interface OneDimension {
+    int move(int index);
+}
