+++ src/main/java/nextstep/ladder/domain/dimension/TwoDimension.java
@@ -1,0 +1,6 @@
+package nextstep.ladder.domain.dimension;
+
+public interface TwoDimension {
+
+    int move(int index);
+}
